split gradle from logic in packages
