+++ buildSrc/src/main/java/io/micronaut/website/docsindex/CategoryFetchImpl.java
@@ -13,8 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.website;
+package io.micronaut.website.docsindex;
 
+import io.micronaut.website.docsindex.Category;
+import io.micronaut.website.docsindex.CategoryFetcher;
+import io.micronaut.website.docsindex.CategoryImpl;
+import io.micronaut.website.docsindex.Repository;
+import io.micronaut.website.docsindex.RepositoryImpl;
+import io.micronaut.website.docsindex.Type;
 import org.gradle.api.tasks.Input;
 import org.yaml.snakeyaml.Yaml;
