+++ buildSrc/src/main/java/io/micronaut/website/docsindex/CategoryImpl.java
@@ -13,7 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.website;
+package io.micronaut.website.docsindex;
+
+import io.micronaut.website.docsindex.Category;
+import io.micronaut.website.docsindex.Repository;
 
 import java.util.List;
