+++ buildSrc/src/main/java/io/micronaut/website/docsindex/IndexRenderer.java
@@ -13,11 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.website;
+package io.micronaut.website.docsindex;
 
-public interface Repository {
-    String getSlug();
-    String getTitle();
-    String getDescription();
-    boolean isSnapshot();
+@FunctionalInterface
+public interface IndexRenderer {
+    String renderAsHtml();
 }
