+++ buildSrc/src/main/java/io/micronaut/website/docsindex/IndexRendererImpl.java
@@ -13,7 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.website;
+package io.micronaut.website.docsindex;
+
+import io.micronaut.website.docsindex.ApiVersionRenderer;
+import io.micronaut.website.docsindex.CategoryFetcher;
+import io.micronaut.website.docsindex.CategoryRenderer;
+import io.micronaut.website.docsindex.DocVersionRenderer;
+import io.micronaut.website.docsindex.IndexRenderer;
+import io.micronaut.website.docsindex.Type;
+import io.micronaut.website.docsindex.Utils;
+import io.micronaut.website.docsindex.VersionsFetcher;
 
 import java.io.IOException;
 import java.io.InputStream;
