+++ buildSrc/src/main/java/io/micronaut/website/docsindex/RepositoryImpl.java
@@ -13,7 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.website;
+package io.micronaut.website.docsindex;
+
+import io.micronaut.website.docsindex.Repository;
 
 public class RepositoryImpl implements Repository {
     private final String slug;
