+++ buildSrc/src/main/java/io/micronaut/website/docsindex/RepositoryRenderImpl.java
@@ -13,7 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.website;
+package io.micronaut.website.docsindex;
+
+import io.micronaut.website.docsindex.Repository;
+import io.micronaut.website.docsindex.RepositoryRenderer;
+import io.micronaut.website.docsindex.Utils;
 
 import java.io.IOException;
 import java.io.InputStream;
