+++ buildSrc/src/main/java/io/micronaut/website/docsindex/Utils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.website;
+package io.micronaut.website.docsindex;
 
 import java.io.BufferedReader;
 import java.io.IOException;
