+++ buildSrc/src/main/java/io/micronaut/website/docsindex/VersionsFetcherImpl.java
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.website;
+package io.micronaut.website.docsindex;
 
+import io.micronaut.website.docsindex.VersionsFetcher;
 import org.yaml.snakeyaml.Yaml;
 
 import java.io.File;
