+++ buildSrc/src/main/java/io/micronaut/website/gradle/MicronautWebsiteDocsIndexPlugin.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.website;
+package io.micronaut.website.gradle;
 
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
